requirements.txt
alembic==1.6.5
aniso8601==9.0.1
apispec==4.7.1
apispec-webframeworks==0.5.2
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
Flask==2.0.1
flask-apispec==0.11.0
Flask-Bcrypt==0.7.1
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
packaging==21.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
webargs==8.0.0
Werkzeug==2.0.1
